implemented getMaxKey/getMinKey using OpenMP
